requirements.txt
selenium == 4.18.1
selenium-wire == 5.1.0
webdriver-manager == 4.0.1
random-password-generator == 2.2.0
Faker == 24.3.0
html2text == 2024.2.26
easyimap == 0.7.0

